+++ components/camel-twitter/src/test/java/org/apache/camel/component/twitter/HomeTimeLineDirectIT.java
@@ -16,36 +16,24 @@
  */
 package org.apache.camel.component.twitter;
 
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * consumes tweets
  */
-public class DirectMessagePollingTest extends CamelTwitterConsumerTestSupport {
-    @Override
-    @BeforeEach
-    public void setUp() throws Exception {
-        super.setUp();
-        /* Uncomment when you need a test direct message
-        TwitterConfiguration properties = new TwitterConfiguration();
-        properties.setConsumerKey(consumerKey);
-        properties.setConsumerSecret(consumerSecret);
-        properties.setAccessToken(accessToken);
-        properties.setAccessTokenSecret(accessTokenSecret);
-        Twitter twitter = properties.getTwitter();
-        twitter.sendDirectMessage(twitter.getScreenName(), "Test Direct Message: " + new Date().toString());
-        */
-    }
+@EnabledIfSystemProperty(named = "enable.twitter.itests", matches = "true",
+                         disabledReason = "Likely has API limits, so it's better to keep it off by default")
+public class HomeTimeLineDirectIT extends CamelTwitterConsumerITSupport {
 
     @Override
     protected String getUri() {
-        return "twitter-directmessage://foo?type=polling&";
+        return "twitter-timeline://home?type=direct&";
     }
 
     @Override
     protected Logger getLogger() {
-        return LoggerFactory.getLogger(DirectMessagePollingTest.class);
+        return LoggerFactory.getLogger(HomeTimeLineDirectIT.class);
     }
 }
