+++ components/camel-twitter/src/test/java/org/apache/camel/component/twitter/UserTimeLineDirectIT.java
@@ -16,20 +16,24 @@
  */
 package org.apache.camel.component.twitter;
 
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * consumes tweets
  */
-public class HomeTimeLinePollingTest extends CamelTwitterConsumerTestSupport {
+@EnabledIfSystemProperty(named = "enable.twitter.itests", matches = "true",
+                         disabledReason = "Likely has API limits, so it's better to keep it off by default")
+public class UserTimeLineDirectIT extends CamelTwitterConsumerITSupport {
+
     @Override
     protected String getUri() {
-        return "twitter-timeline://home?type=polling&";
+        return "twitter-timeline://user?type=direct&user=brettemeyer&";
     }
 
     @Override
     protected Logger getLogger() {
-        return LoggerFactory.getLogger(HomeTimeLinePollingTest.class);
+        return LoggerFactory.getLogger(UserTimeLineDirectIT.class);
     }
 }
